fixed warn about sign
git-svn-id: https://btanks.svn.sourceforge.net/svnroot/btanks/trunk@2526 5c0bf991-d514-0410-a65f-ce4832e571f3

+++ menu/map_picker.cpp
@@ -158,7 +158,7 @@ void MapPicker::fillSlots() const {
 	MenuConfig->fill(_maps[_index].name, _picker->getVariant(), config);
 	int idx1 = -1, idx2 = -1;
 	for(size_t i = 0; i < config.size(); ++i) {
-		if (i >= _maps[_index].slots)
+		if (i >= (size_t)_maps[_index].slots)
 			break;
 		
 		PlayerSlot &slot = PlayerManager->getSlot(i);
